refactor(index.tsx): moved the counter code into a new component along with relevant states and handlers
